move argparse functions to conda_argparse
