+++ conda/cli/main_remove.py
@@ -12,12 +12,13 @@ import logging
 from os.path import abspath, join
 import sys
 
-from .common import (add_parser_channels, add_parser_help, add_parser_json,
-                     add_parser_no_pin, add_parser_no_use_index_cache, add_parser_offline,
-                     add_parser_prefix, add_parser_pscheck, add_parser_quiet,
-                     add_parser_use_index_cache, add_parser_use_local, add_parser_yes, confirm_yn,
+from .common import (confirm_yn,
                      ensure_override_channels_requires_channel, ensure_use_local, names_in_specs,
                      specs_from_args, stdout_json)
+from .conda_argparse import (add_parser_channels, add_parser_help, add_parser_json,
+                             add_parser_no_pin, add_parser_no_use_index_cache, add_parser_offline,
+                             add_parser_prefix, add_parser_pscheck, add_parser_quiet,
+                             add_parser_use_index_cache, add_parser_use_local, add_parser_yes)
 
 try:
     from cytoolz.itertoolz import groupby
