+++ conda_env/cli/main_attach.py
@@ -1,9 +1,11 @@
 from argparse import RawDescriptionHelpFormatter
-from ..utils.notebooks import Notebook
-from conda.cli import common
+
+from conda.cli.conda_argparse import add_parser_json
+
+from .common import get_prefix
 from ..env import from_environment
-# conda env import
-from conda_env.cli.common import get_prefix
+from ..utils.notebooks import Notebook
+
 description = """
 Embeds information describing your conda environment
 into the notebook metadata
@@ -63,7 +65,7 @@ def configure_parser(sub_parsers):
         action='store',
         default=None
     )
-    common.add_parser_json(p)
+    add_parser_json(p)
     p.set_defaults(func=execute)
